% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoLog10.R
\name{pseudoLog10}
\alias{pseudoLog10}
\title{Return the pseudo logarithm, base 10.}
\usage{
pseudoLog10(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
pseudo logarithm, base 10 of x
}
\description{
Return the pseudo logarithm (base 10) of x, which is close to
sign(x)*log10(abs(x)) for x such that abs(x) is large
and doesn't "blow up" near zero.  Useful
for transforming wide-range variables that may be negative
(like profit/loss).
}
\details{
See: \url{http://www.win-vector.com/blog/2012/03/modeling-trick-the-signed-pseudo-logarithm/}

NB: This transform has the undesirable property of making most
signed distributions appear bi-modal around the origin, no matter
what the underlying distribution really looks like.
The argument x is assumed be numeric and can be a vector.
}
\examples{

pseudoLog10(c(-5, 0, 5))
# should be: [1] -0.7153834  0.0000000  0.7153834

}
